Fix service hash logic
The big bug was that we weren't hashing the pinned service image. This
lead to hashes changing when an image's docker tag had change but the
underlying sha256 hash had not.

In addition, we were hashing an internal compose-ref data struct that
we weren't saving anywhere. This meant the hash wasn't reproducible.

The new logic is now reproducible by taking the bytes of a service
section (plus a '\n' byte) and doing a sha256sum

Signed-off-by: Andy Doan <andy@foundries.io>